APPLY WORKLOAD MANAGER TO GCPS AND GCSimpl. CHANGE CUTTING METRIC TRIANGLES CONDITION TO EQUAL LENGTHS.
We apply workload manager to GCPS::nextPrefSufIteration and GCSimpl::removeShortParallelPaths, since there may be unbalanced work done. We try to make it better balanced, hence improving performance.

Changed condition in cutting metric triangles to equal length - an edge (a,c) will be removed only if there exist edges (a,b) and (b,c) such that sum of their weights is equal to the weight of (a,c). This reduces indels and mismatches.