Fix Bitset::mismatch() function. Fix mismatch range check in GCPS. Change prefixBuckets to power of 2.
There was an error in the mismatch function ( __builtin_ctz(V[i]) and should be __builtin_ctz(V[i]^otg.V[i]) ).

In GCPS there was a condition that first mismatch may occur at position rA->size() - offsetDiff, but nucleotides are enocded on two bits, so it should be two times that value.

Graph::writeBasicStatistics() was changed to do nothing (just comment the early return statement to make it work).

prefixBuckets is now a power of 2 - modulo operations are a bit faster now.